+++ tests/hud_meters.rs
@@ -15,8 +15,8 @@ fn split_meter_value_returns_full_blocks_and_partial_block() {
         game::split_meter_value(&mut engine, &mut r);
 
         assert_eq!(engine.state.scratch0(), expected_partial, "value {value}");
-        assert_eq!(r.value, expected_partial, "value {value}");
-        assert_eq!(r.offset, expected_blocks, "value {value}");
+        assert_eq!(r.value as i32, expected_partial, "value {value}");
+        assert_eq!(r.offset as i32, expected_blocks, "value {value}");
     }
 }
